restructured grouped topk kernel
